merge: update branch with latest dev

+++ packages/central-server/src/social/getRewardsForUser.js
@@ -1,11 +1,6 @@
-/**
- * Tupaia
- * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
- */
-
 export const getRewardsForUser = async (database, userId, projectId = '') => {
   const [rewards] = await database.executeSql(
-    `SELECT user_id, COUNT(*) as coconuts, FLOOR(COUNT(*) / 100) as pigs
+    `SELECT user_id, COUNT(*)::int as coconuts, FLOOR(COUNT(*) / 100)::int as pigs
     FROM survey_response
     JOIN survey on survey.id=survey_id
     WHERE user_id = ?
